requirements.txt
streamlit
pymupdf
pandas
pillow
tqdm
python-dotenv
google-genai